fix: 采购单数量改为BigDecimal精确小数，修复下单采购量与预览不一致(含lossRate)

+++ backend/src/main/java/com/fashion/supplychain/finance/orchestration/MaterialReconciliationOrchestrator.java
@@ -607,7 +607,7 @@ public class MaterialReconciliationOrchestrator {
             return 0;
         }
         int aq = purchase.getArrivedQuantity() == null ? 0 : purchase.getArrivedQuantity();
-        int pq = purchase.getPurchaseQuantity() == null ? 0 : purchase.getPurchaseQuantity();
+        int pq = purchase.getPurchaseQuantity() == null ? 0 : purchase.getPurchaseQuantity().intValue();
         if (pq > 0) {
             try {
                 return Math.max(0, materialPurchaseService.computeEffectiveArrivedQuantity(pq, aq));
